+++ src/svd/register.rs
@@ -8,6 +8,7 @@ use crate::types::Parse;
 use crate::elementext::ElementExt;
 #[cfg(feature = "unproven")]
 use crate::encode::Encode;
+use crate::error::*;
 use crate::svd::{dimelement::DimElement, registerinfo::RegisterInfo};
 use anyhow::Result;
 
@@ -40,7 +41,7 @@ impl Parse for Register {
 
         if tree.get_child("dimIncrement").is_some() {
             let array_info = DimElement::parse(tree)?;
-            assert!(info.name.contains("%s"));
+            check_has_placeholder(&info.name, "register")?;
             if let Some(indices) = &array_info.dim_index {
                 assert_eq!(array_info.dim as usize, indices.len())
             }